@@ -1,6 +1,6 @@
 {
   "name": "earthquakeph",
-  "version": "5.1.0",
+  "version": "5.1.1",
   "private": true,
   "dependencies": {
     "@lottiefiles/react-lottie-player": "^3.6.0",
@@ -11,7 +11,7 @@
     "chart.js": "^4.4.8",
     "chartjs-plugin-deferred": "^2.0.0",
     "mapbox-gl": "^3.10.0",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-chartjs-2": "^5.3.0",
     "react-dom": "^18.3.1",
     "react-scripts": "^5.0.1",
@@ -47,7 +47,7 @@
     "dependency-cruiser": "^16.10.0",
     "jest-watch-typeahead": "^0.6.5",
     "stylelint": "^16.15.0",
-    "stylelint-config-standard": "^34.0.0",
+    "stylelint-config-standard": "^37.0.0",
     "stylelint-config-standard-scss": "^14.0.0"
   },
   "jest": {
