@@ -15,15 +15,16 @@ Developer's Website: https://jpvitan.com/
 
 import App from './App'
 import React from 'react'
-import ReactDOM from 'react-dom'
+import ReactDOM from 'react-dom/client'
 import * as serviceWorkerRegistration from './serviceWorkerRegistration'
 import './index.scss'
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'))
+
+root.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 )
 
 serviceWorkerRegistration.register()
